backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
motor>=3.3.2
pymongo>=4.8.0
python-jose[cryptography]==3.3.0
bcrypt>=4.1.0
python-multipart>=0.0.9
transformers>=4.40.0
torch>=2.2.0
pydantic[email]>=2.9.0
pydantic-settings>=2.3.0
pytest>=8.0.0
httpx>=0.27.0
pytest-asyncio>=0.23.0
gunicorn>=21.2.0
feedparser>=6.0.10
